Add LLVM based JIT.
For each supported instruction create a function that
implements this instruction. Clang is used to compile these
functions to LLVM bitcode. The simulator takes sequences of
instructions and uses the LLVM C API to build a function
calling the functions corresponding to the instructions in the
sequence. This is then JIT compiled.