Refactor tracing to avoid a second buffer for events / exceptions / etc.
Delay printing of the start of the instruction line until the first
register write if there is one or until the instruction has completed.
This ensures that when we get an event / exception / interupt we won't be
in the middle of printing a line so we can output trace information
straight away without an extra buffer.