Don't rely on find_package to find LLVM.
find_package only works if LLVM was built with cmake and this isn't the
case with LLVM binaries shipped with various Linux distributions (e.g.
Ubuntu). This is based on the FindLLVM cmake module that was previously
removed in commit b4665b72f261f78ebfc14abae66f60f178ed400b.